+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## 0.18.9
+### Added
+- Endpoint to locate a block by unix timestamp.
+
 ### Removed
 - No more persistence for peers due to dependency on newest haskoin-node.